+++ MefistoDB/Tables/Teams.sql
@@ -2,6 +2,6 @@
 (
 	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [Name] NVARCHAR(MAX) NOT NULL, 
-    [Odds] FLOAT NOT NULL, 
+    [Odds] DECIMAL(5, 2) NOT NULL, 
     [CoversWinPercentage] INT NULL
 )
